### 2022-09-07 #4743 @editorialbot
[PRE REVIEW]: PYroMat: A Python package for thermodynamic properties
**Submitting author:** @jranalli (<a href="http://orcid.org/0000-0002-8184-9895">Joseph Ranalli</a>)
**Repository:** https://github.com/chmarti1/PYroMat
**Branch with paper.md** (empty if default branch): joss2022
**Version:** 2.2.1
