### 2022-09-07 #4743 @kthyng
Hi @jranalli! The paper file in your `joss2022` branch needs to be needs "paper.md" for the bot to find it.
