### 2022-09-07 #4743 @jranalli
> In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission

From their area descriptions, I identified a few potential matches in no particular order: ronshnapp, mefuller, zadehsey, erasmuss, espottesmith, fwitte, mwerwee
