### 2022-09-10 #4743 @jgostick
This one is my jam!  I will almost certainly use this in my work.
