### 2022-09-12 #4757 @editorialbot
[REVIEW]: PYroMat: A Python package for thermodynamic properties
**Submitting author:** @jranalli (<a href="http://orcid.org/0000-0002-8184-9895">Joseph Ranalli</a>)
**Repository:** https://github.com/chmarti1/PYroMat
**Branch with paper.md** (empty if default branch): joss2022
**Version:** 2.2.4
