### 2022-09-16 012eee2
Modifications in response to issue #52 to go back to original call signatures for deprecated T_s, etc. function calls.
