+++ src/pyromat/registry/ig2.py
@@ -1249,18 +1249,23 @@ T, p, d, v, e, h, and s.
         """Temperature as a function of entropy
 ** Depreciated - use T() **
         
-    T = T_s(s)
+   T = T_s(s)
         or
     T = T_s(s,p)
-        or
-    T = T_s(s,d)
 
 Accepts unit_energy / unit_matter / unit_temperature
         unit_pressure
         unit_matter / unit_volume
 Returns unit_temperature
 """
-        return self.T(*varg, s=s, **kwarg)
+        if len(varg) > 0:
+            if 'p' in kwarg:
+                raise pm.utility.PMParamError('p was specified both positionally and with a keyword.')
+            kwarg['p'] = varg[0]
+        if len(varg) > 1:
+            raise pm.utility.PMParamError('There are only two positional arguments: s, p.')
+
+        return self.T(s=s, **kwarg)
 
 
     def T_h(self,h,*varg,**kwarg):
@@ -1279,8 +1284,16 @@ Accepts unit_energy / unit_matter / unit_temperature
         unit_pressure
 Returns unit_temperature
 """
-        # Convert the 
-        return self.T(*varg, h=h, **kwarg)
+        if len(varg) > 0:
+            if 'p' in kwarg:
+                raise pm.utility.PMParamError(
+                    'p was specified both positionally and with a keyword.')
+            kwarg['p'] = varg[0]
+        if len(varg) > 1:
+            raise pm.utility.PMParamError(
+                'There are only two positional arguments: h, p.')
+
+        return self.T(h=h, **kwarg)
 
 
     def p_s(self,s,*varg, **kwarg):
