### 2022-09-22 chmarti1/PYroMat#54 @espottesmith
[Review] LICENSE file does not contain actual license text
Very small issue: License.txt makes reference to the GPLv3 license but does not actually reproduce the text. For acceptance into JOSS, the contents of the license should actually be provided (see [here](https://joss.readthedocs.io/en/latest/review_criteria.html)).
