### 2022-09-22 4a60d4c
Added full text of GPL3
