### 2022-09-22 #4757 @espottesmith
My review is complete.

The authors have produced and described `pyromat`, a utility for calculation of gas-phase (and, to a certain extent, mixed gas/liquid) thermodynamic properties, based on established models. They provide data for ~1000 unique species, which include both common and uncommon gas-phase molecules and ions. The code is written clearly, and the documentation is, for the most part, quite good. While the use case for `pyromat` in research seems unclear to me (note: I do not study gas-phase or combustion thermochemistry), this seems an extremely necessary tool for students of chemistry/chemical engineering, as well as industrial engineers. My comments are below.

Compliments:
- It seems that, provided the user has some data available, this code is (relatively) easily extensible. It also appears to be relatively easily configurable, though better documentation on all of the configuration parameters could perhaps be desired.

Minor issues:
- The LICENSE.txt file refers to the GPL but does not reproduce the text of the license. As I understand JOSS' guidelines, this is not acceptable. I have created an issue (https://github.com/chmarti1/PYroMat/issues/54) regarding this.
- One of the examples in the README does not work, though it does with small modifications (see https://github.com/chmarti1/PYroMat/issues/55)
- In the text of the paper, references should probably be made to the models used to populate the data in `pyromat`.
- There are no guidelines on how an external developer should contribute to the code.

Major issues:
- There are areas where the code feels incomplete. Most critically, there is not a test suite. Validation - ensuring that you obtain expected values for a small set of materials, or ensuring that the predicted values are close to experimental values for well-studied substances like water - is essential to give the user confidence in the reported results.
- Some features also more generally feel half-finished; in `apps/cycle.py`, for instance, the RefrigerationCycle, OttoCycle, and DieselCycle classes are defined but not implemented.
