### 2022-09-22 #4757 @jranalli
Thank you very much for your time in the review. I'd just like to clarify the two major issues:
- We have some tests stored in branch `tests-jranalli`. Specifically the files `test_ig` and `test_mp1`. Is there some type of documentation of that we could add that would help to satisfy this concern?
- I agree that the cycle codes are half finished. It represents a longer term goal for the project to include some more holistic calculations/educational modules. The part of the code that we're really submitting is the primary property calculation portion. Is there a way that documenting that would satisfy the concern, or would it be necessary to remove those from the library.
