### 2022-09-26 370f0b8
Remove old "apps" codes as cleanup for josspaper. Retained in extra branch.
