### 2022-09-26 8d67fcd
Add two basic tests.
