+++ src/test/test_ig.py
@@ -1,0 +1,565 @@
+import pyromat as pm
+import numpy as np
+from pytest import approx, raises
+import pytest
+
+
+class TestInputErrors:
+    @pytest.fixture(params=('ig.O2', 'ig.BH3O3', 'ig.air'),
+                    ids=('oxygen-ig2', 'BH3O3-ig', 'air-ig'))
+    def gas(self, request):
+        o = pm.get(request.param)
+        return o
+
+    def test_T_only_varg(self, gas):
+        # Based on how .T is written, it interprets this as a temperature given
+        # So just make sure that works
+        assert gas.T(300) == gas.T(T=300, p=pm.config['def_p'])
+
+    def test_T_double_specified(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(300, T=300)
+
+    def test_p_double_specified(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(300, 1, p=1)
+
+    def test_varg_specified(self, gas):
+        assert gas.s(300, 1) == gas.s(p=1, T=300)
+
+    def test_varg_toomany(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(300, 1, 5)
+
+    def test_any_arg_toomany(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(300, 1, s=5)
+        with raises(pm.utility.PMParamError):
+            gas.T(T=300, p=1, s=5)
+
+    def test_default_usage(self, gas):
+        assert gas.s() == gas.s(T=pm.config['def_T'], p=pm.config['def_p'])
+        assert gas.s(T=pm.config['def_T']) == gas.s(T=pm.config['def_T'], p=pm.config['def_p'])
+        assert gas.s(p=pm.config['def_p']) == gas.s(T=pm.config['def_T'], p=pm.config['def_p'])
+
+    def test_invarg_toomany(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(h=300, s=5)
+        with raises(pm.utility.PMParamError):
+            gas.T(e=300, s=5)
+
+    def test_d_v_collision(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(d=1, v=1)
+
+    def test_illegal_prop(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(T=300, fakeprop=10000)
+
+    def test_T_oob_all(self, gas):
+        with raises(pm.utility.PMParamError):
+            gas.T(T=30000, p=1)
+
+    def test_T_oob_some(self, gas):
+        vals = gas.T(T=[300, 30000], p=1)
+        assert vals[0] == 300
+        assert np.isnan(vals[1])
+
+
+class TestGeneral:
+
+    @pytest.mark.parametrize('sub', ('ig.O2', 'ig.BH3O3', 'ig.air'), ids=('oxygen-ig2', 'boric acid-ig', 'air-ig'))
+    def test_get(self, sub):
+        # Just confirm no error
+        igobj = pm.get(sub)
+
+    @ pytest.mark.parametrize('gas, expect', [('ig.O2',(200, 6000)), ('ig.BH3O3',(298, 6000)), ('ig.air',(200, 6000))])
+    def test_Tlim(self, gas, expect):
+        assert pm.get(gas).Tlim() == approx(expect)
+
+    @ pytest.mark.parametrize('gas, expect', [('ig.O2', 31.9988), ('ig.BH3O3', 61.833), ('ig.air', 28.96493)])
+    def test_mw(self, gas, expect):
[…]
