+++ src/test/test_mp1.py
@@ -1,0 +1,875 @@
+import pyromat as pm
+import numpy as np
+from pytest import approx, raises
+import pytest
+
+
+class TestInputErrors:
+    @pytest.fixture(params=('mp.H2O', 'mp.C2H2F4'),
+                    ids=('water', 'R134a'))
+    def subst(self, request):
+        o = pm.get(request.param)
+        return o
+
+    def test_T_only_varg(self, subst):
+        # Based on how .T is written, it interprets this as a temperature given
+        # So just make sure that works
+        assert subst.T(300) == subst.T(T=300, p=pm.config['def_p'])
+
+    def test_T_double_specified(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(300, T=300)
+
+    def test_p_double_specified(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(300, 1, p=1)
+
+    def test_varg_specified(self, subst):
+        assert subst.s(300, 1) == subst.s(p=1, T=300)
+
+    def test_varg_toomany(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(300, 1, 5)
+
+    def test_any_arg_toomany(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(300, 1, s=5)
+        with raises(pm.utility.PMParamError):
+            subst.T(T=300, p=1, s=5)
+
+    def test_default_usage(self, subst):
+        assert subst.s() == subst.s(T=pm.config['def_T'], p=pm.config['def_p'])
+        assert subst.s(T=pm.config['def_T']) == subst.s(T=pm.config['def_T'], p=pm.config['def_p'])
+        assert subst.s(p=pm.config['def_p']) == subst.s(T=pm.config['def_T'], p=pm.config['def_p'])
+        # And for saturated
+        assert subst.ss() == subst.ss(T=pm.config['def_T'])
+        assert subst.Ts() == subst.Ts(p=pm.config['def_p'])
+        assert subst.ps() == subst.ps(T=pm.config['def_T'])
+
+    def test_invarg_toomany(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(h=300, s=5)
+        with raises(pm.utility.PMParamError):
+            subst.T(e=300, s=5)
+
+    def test_d_v_collision(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(d=1, v=1)
+
+    def test_illegal_prop(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(T=300, fakeprop=10000)
+
+    def test_T_oob_all(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(T=30000, p=1)
+
+    def test_T_oob_some(self, subst):
+        vals = subst.T(T=[300, 30000], p=1)
+        assert vals[0] == 300
+        assert np.isnan(vals[1])
+
+    def test_p_oob_all(self, subst):
+        with raises(pm.utility.PMParamError):
+            subst.T(T=300, p=1e100)
+
+    def test_p_oob_some(self, subst):
+        vals = subst.p(T=300, p=[1, 1e100])
+        assert vals[0] == approx(1)
+        assert np.isnan(vals[1])
+
[…]
