### 2022-09-26 d78b062
Add the CI github action
