+++ .github/workflows/test.yml
@@ -1,0 +1,30 @@
+name: test
+
+on:
+  push:
+    branches: [ "addtests-joss" ]
+  pull_request:
+    branches: [ "addtests-joss" ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install pytest
+        python setup.py install
+    - name: Test Ideal Gases
+      working-directory: ./src/test
+      run: pytest test_ig.py
+    - name: Test Multiphase
+      working-directory: ./src/test
+      run: pytest test_mp1.py
