### 2022-09-27 #4757 @chmarti1
Thanks very much for your work.  I'll make a few comments here:

**With Respect to Test Suites**
This idea really needs to be divided into two parts: (1) the testing required to validate a new data set that is imported into the suite, and (2) testing required to validate new class functionality or new features.

(1) New data are carefully validated against reference data provided by the same sources from which the models are entered.  This is a time-consuming process - for multi-phase source data, it has to be done manually for every data source, since the raw data are being pulled from publications.  However, once validated, the work is complete, and does not need to be repeated with new releases, so no automation suite is used.

We deliberately do not currently publish validation data.  In older versions of the code, I used to include records of my validation work, but I have received emails from private companies that were attempting to interpret this as a warranty of some kind.  My official policy is that the sources of the original data are cited, and users should always validate their calculations against their own experimental data in the context of their application.

However, the codes that were used to validate the original ideal gas model data are actually still coded directly into the classes.  The `ig` class has a method called `_ig__test()`, and `ig2` has `_test()`.  The `ig` substances even have validation data tables embedded directly in their raw data.  `_ig__test()` now complains about errors because it has not been updated along with major version changes, but it still successfully validates the original source data.  This has even been used to demonstrate continuity errors in NIST's source data.

(2) On this front, @jranalli deserves credit for making similar arguments for some time.  In brief, this is a proposal that makes sense, but to implement it well, it would require an effort at least as difficult as writing the package itself.  Now that @jranalli is contributing so heavily to the project, it may make sense to include his validation suites in the main branch.

The challenge is a familiar one: the number of permutations of input/output possibilities is now so large (e.g. array/scalar, liquid/mixed/vapor/super-critical, primary/inverse properties, near limits/centered, countless permutations of units, and all of the combinations of those), that historically, hard-coded test suites have not performed well enough to justify the effort.  So far, our most successful model has been to run ad-hoc targeted tests to try to break the new functionality, then @jranalli tries harder with his own test suite, then we release to the community.  This workflow has functioned pretty well so far, but we are always happy to improve.

**Undocumented Partially Coded Modules**
True - we do have some undocumented code in the master branch.  It is not currently part of package functionality.  The `apps` module is utterly undocumented (except in-line) and is not imported by default.  That module actually serves a specific (if slightly secret) purpose for our efforts to port to a web-based interface for PYroMat.  The code there is part of a (still live) proof-of-concept portion of the [website](http://pyromat.org).  It will be cleaned up in later releases, but we have decided to leave it as an easter egg for any interested users who may want to do their own developing.

**LICENSE**
True.  That will be easily corrected.

**README failure**
This is actually a change in the functionality of `info()` in the last release.  The new behavior was documented on the [website](http://pyromat.org/features.html), and in the [handbook](http://pyromat.org/pdf/handbook.pdf) (see sections 2.2.3 and 2.2.4), but I forgot to update the README.  Oops.  Good catch!
