### 2022-09-27 chmarti1/PYroMat#54 @chmarti1
Quite right.  This was the way I originally assembled the package back in 2015, and I haven't really revisited it since.  Easily fixed.
