+++ src/test/test_gen_conditions.py
@@ -1,0 +1,189 @@
+import numpy as np
+import pytest
+from pytest import approx, raises
+
+import pyromat as pm
+
+
+# Note that this generates an extremely large number of tests, rougly 800k
+# for ig. So we're currently running things for a single PM class at a time.
+# We should look for ways to manage this or separate things into different
+# runs.
+
+# The purpose of these tests is really to hit every single substance to make
+# sure that our calculations are reliable across all, rather than the few
+# targeted substances listed previously.
+
+n = 5  # Number of points per temperature/pressure range
+pmclass = 'mp1'  # The class to test: one of igmix, ig, ig2, mp1
+
+
+# Which parameters to test the outputs of based on class
+params = {'mp1': ('p', 'T', 'd', 'v', 'e', 'h', 's', 'x'),
+          'ig': ('p', 'T', 'd', 'v', 'e', 'h', 's'),
+          'ig2': ('p', 'T', 'd', 'v', 'e', 'h', 's'),
+          'igmix': ('p', 'T', 'd', 'v', 'e', 'h', 's')
+          }
+
+
+def getT(sub, i, n):
+    """
+    Get a temperature from the standard list by index using a uniform
+    methodology by substance. This is necessary because each substance has its
+    own legal temperatures, but the pytest.fixtures require that everything
+    be prespecified. Thus we feed the fixtures a list of indices, and allow
+    this function to generate the actual data.
+    """
+    if sub.data['id'].startswith('mp'):
+        Tmin, Tmax = sub.Tlim()
+        # Tt, pt = sub.triple()
+        # Tmin = max(Tmin, Tt)
+        Tmax *= 0.999
+        Tmin *= 1.001
+        Tlist = np.linspace(Tmin, Tmax, n)
+    else:
+        Tmin, Tmax = sub.Tlim()
+        Tmin *= 1.001
+        Tmax *= 0.999
+        Tlist = np.linspace(Tmin, Tmax, n)
+    return Tlist[i]
+
+
+def getp(sub, i, n):
+    """
+    Get a pressure from the standard list by index using a uniform
+    methodology by substance. This is necessary because each substance has its
+    own legal temperatures, but the pytest.fixtures require that everything
+    be prespecified. Thus we feed the fixtures a list of indices, and allow
+    this function to generate the actual data.
+    """
+    if sub.data['id'].startswith('mp'):
+        pmin, pmax = sub.plim()
+        # at low T and high p, we can actually cause density to go out of range
+        pmax2 = sub.p(d=sub.data['dlim'][1], T=sub.data['Tlim'][0])
+        pmax = min(pmax, pmax2)
+        pmax *= 0.999
+        Tt, pt = sub.triple()
+        pt *= 1.001
+        plist = np.linspace(pt, pmax, n)
+    else:
+        Tmin, Tmax = sub.Tlim()
+        pmin, pmax = np.array([sub.p(T=Tmin, d=0.01), sub.p(T=Tmax, d=1000)]).flatten()
+        plist = np.linspace(pmin, pmax, n)
+    return plist[i]
+
+
+class TestArbitary:
+
+    @pytest.fixture(params=[item.data['id'] for item in pm.search(pmclass=pmclass)], ids=[item.data['id'] for item in pm.search(pmclass=pmclass)])
+    def sub(self, request):
+        sub = pm.get(request.param)
+        return sub
+
[…]
