+++ src/pyromat/registry/ig2.py
@@ -1245,7 +1245,7 @@ T, p, d, v, e, h, and s.
         return out
 
 
-    def T_s(self,s,*varg, **kwarg):
+    def T_s(self,s,p=None, **kwarg):
         """Temperature as a function of entropy
 ** Depreciated - use T() **
         
@@ -1258,17 +1258,12 @@ Accepts unit_energy / unit_matter / unit_temperature
         unit_matter / unit_volume
 Returns unit_temperature
 """
-        if len(varg) > 0:
-            if 'p' in kwarg:
-                raise pm.utility.PMParamError('p was specified both positionally and with a keyword.')
-            kwarg['p'] = varg[0]
-        if len(varg) > 1:
-            raise pm.utility.PMParamError('There are only two positional arguments: s, p.')
-
+        if p is not None:
+            return self.T(s=s, p=p)
         return self.T(s=s, **kwarg)
 
 
-    def T_h(self,h,*varg,**kwarg):
+    def T_h(self,h,p=None,**kwarg):
         """Temperature as a function of enthalpy
 ** Depreciated - use T() **
 
@@ -1284,15 +1279,8 @@ Accepts unit_energy / unit_matter / unit_temperature
         unit_pressure
 Returns unit_temperature
 """
-        if len(varg) > 0:
-            if 'p' in kwarg:
-                raise pm.utility.PMParamError(
-                    'p was specified both positionally and with a keyword.')
-            kwarg['p'] = varg[0]
-        if len(varg) > 1:
-            raise pm.utility.PMParamError(
-                'There are only two positional arguments: h, p.')
-
+        if p is not None:
+            return self.T(h=h, p=p)
         return self.T(h=h, **kwarg)
 
 
