+++ joss2022/paper.bib
@@ -62,3 +62,41 @@
 }
 
 
+@ARTICLE{span:2003:1,
+  author = {R. Span and W. Wagner},
+  title = {Equations of state for technical applications. I. Simultaneously optimized functional forms for nonpolar and polar fluids},
+  journal = {International Journal of Thermophysics},
+  volume = {24},
+  number = {1},
+  year = {2003},
+  pages = {1-39},
+}
+@ARTICLE{span:2003:2,
+  author = {R. Span and W. Wagner},
+  title = {Equations of state for technical applications. II. Results for nonpolar fluids},
+  journal = {International Journal of Thermophysics},
+  volume = {24},
+  number = {1},
+  year = {2003},
+  pages = {41-109},
+}
+@ARTICLE{span:2003:3,
+  author = {R. Span and W. Wagner},
+  title = {Equations of state for technical applications. III. Results for polar fluids},
+  journal = {International Journal of Thermophysics},
+  volume = {24},
+  number = {1},
+  year = {2003},
+  pages = {111-162},
+}
+@MISC{nist:webbook,
+  howpublished = {National Institute for Standards and Technology, https://webbook.nist.gov/},
+  note = {accessed: Jun, 2021},
+}
+@TECHREPORT{mcbride:1993,
+  author = {Bonnie J. {McBride} and Sanford Gordon and Martin A. Reno},
+  title = {Coefficients for Calculating Thermodynamic and Transport Properties of Individual Species},
+  institution = {NASA},
+  year = {1993},
+  number = {4513},
+}
