+++ joss2022/paper.md
@@ -58,7 +58,7 @@ array([386.32246521])
 
 `PYroMat` can be configured to work in virtually any system of units in common use, and users can specify states with almost any pair of available properties.  The interface is well documented using Python's in-line help system, on the PYroMat website, [pyromat.org](http://pyromat.org), and in the regularly updated PYroMat User and Developer's Handbook [@pmhandbook].
 
-Substances are organized into two collections: ideal gas (`ig`) and multi-phase (`mp`).  They are further organized by the specific underlying data model (class), so users always know what assumptions are implicit in the property calculations.  All of the classes implement a common flexible call signature, so the user can concentrate on the application instead of the nuances of the model.  As of version 2.2.1, `PYroMat` implements ideal gas data using the Shomate model (used by NIST) with the `ig` class, the time tested `NASA` polynomials with the `ig2` class, and ideal gas mixtures using the `igmix` class.  The so-called Span and Wagner multi-phase models are implemented with the `mp1` class.  There are plans to add more classes in later releases.
+Substances are organized into two collections: ideal gas (`ig`) and multi-phase (`mp`).  They are further organized by the specific underlying data model (class), so users always know what assumptions are implicit in the property calculations.  All of the classes implement a common flexible call signature, so the user can concentrate on the application instead of the nuances of the model.  As of version 2.2.1, `PYroMat` implements ideal gas data using the Shomate model (used by NIST) with the `ig` class [@nist:webbook], the time tested `NASA` polynomials with the `ig2` class [@mcbride:1993], and ideal gas mixtures using the `igmix` class [@pmhandbook].  The so-called Span and Wagner multi-phase models are implemented with the `mp1` class [@span:2003:1; @span:2003:2; @span:2003:3].  There are plans to add more classes in later releases.
 
 Users can conveniently browse for substances that interest them by name, chemical formula, atomic composition, molecular weight, InChIE or CAS identifier string, `PYroMat` class, or by collection.  For example, this search returns a table of all substances that contain two nitrogen atoms.
 
