### 2022-09-27 #4757 @jranalli
Fixes implemented in the repository:
- [License file extended](https://github.com/chmarti1/PYroMat/pull/56)
- [README demo corrected](https://github.com/chmarti1/PYroMat/pull/57)
- [Continuous Integration for functionality tests added](https://github.com/chmarti1/PYroMat/pull/59)
- [Moved unfinished code (i.e. `apps` folder) to a separate branch to eliminate confusion](https://github.com/chmarti1/PYroMat/pull/58).
- Citations added to mention of models in text of the paper (see previous post for updated PDF)
