### 2022-10-08 chmarti1/PYroMat#60 @fwitte
[Review] Suggestion for the paper and a few typos I found going through the online documentation
Hi,

# Paper

- https://github.com/chmarti1/PYroMat/blob/ffe8f4508a19c182fc147d687063569e9816987f/joss2022/paper.md?plain=1#L101
  I suggest you are either a little bit more specific about the system (`single core of an old laptop`) or just remove that last part of the sentence.

# Online Documentation

- See https://github.com/chmarti1/pyrodoc/issues/4
