### 2022-10-08 chmarti1/PYroMat#61 @fwitte
[Review] Default unit converters for time are incorrect in subsecond range
Following the logic of the length converters https://github.com/chmarti1/PYroMat/blob/d65cb2c73bd14b756deca4b7271246322515d221/src/pyromat/units.py#L277-L281, the time converts for subsecond time should be with a negative power, or?

https://github.com/chmarti1/PYroMat/blob/d65cb2c73bd14b756deca4b7271246322515d221/src/pyromat/units.py#L292-L294

I unfortunately did not yet find a way to test it myself, that's why I opened the issue.
