### 2022-10-08 chmarti1/PYroMat#61 @fwitte
So, I cannot find where the "default" time conversion table is used. But calling something like `time.__call__(10, "s", "ms")` after the variable definition in the `setup()` function does indeed return `0.01` instead of the expected `10000`.
