### 2022-10-08 chmarti1/PYroMat#62 @fwitte
[Review] Insert citation to NumPy in Paper
Hi @jranalli and @chmarti1,

since the package depends on NumPy, you could insert a citation here, see: https://numpy.org/citing-numpy/.

https://github.com/chmarti1/PYroMat/blob/ffe8f4508a19c182fc147d687063569e9816987f/joss2022/paper.md?plain=1#L101
