### 2022-10-08 chmarti1/PYroMat#61 @fwitte
Furthermore, the example conversion inch to feet is not correct:
