### 2022-10-10 chmarti1/PYroMat#61 @jranalli
Thanks for these! I wrote a unit testing script to validate these as the fix was implemented, and identified a few other bugs as well.

@chmarti1 I'll create a separate branch w/pull request when finished, for you to review.
