### 2022-10-10 chmarti1/PYroMat#61 @chmarti1
Thanks for your diligence!

1. **time units:** You are correct - this behavior is wrong, and it should be corrected.  The good news is that it does not affect the package's functionality - none of the properties use time.  It was merely added for completeness.  The conversion values are set on line 291 inside the `setup()` function.  They are commented as "validated" in 2017.  I could trace my records to see why they are not correct, but it is easier to just fix them.

2. **length units:** I don't see an issue with length conversion.  Here are the results I get, which are correct.
```python
>>> pm.units.length(6, from_units='in', to_units='ft')
0.5
>>> pm.units.length(1,from_units='m', to_units='in')
39.37007874015748
```

3. **Tstd** The `setup()` function assigns values to all of the constants (including `const_Tstd`) and then uses those to construct the unit conversion tables.  It appears you are encountering an error on line 263, shortly after `const_Tstd` is assigned on line 259.  The `Tstd` argument appears first in the `setup()` call signature, so if the setup function were being called as if it were a bound method (instead of a function), that might cause the `units` module to be passed to `Tstd` instead of a number.

When I load PYroMat normally,
```
>>> import pyromat as pm
>>> pm.units.const_Tstd
273.15
>>> pm.units.const_dstd
44.61503340547032
>>> pm.units.setup(Tstd=298.15)
>>> pm.units.const_Tstd
298.15
>>> pm.units.const_dstd
40.87404452357611
```
