### 2022-10-10 c55747f
Add tests for the unit conversions
