+++ src/test/test_units.py
@@ -1,0 +1,322 @@
+import pyromat as pm
+import pytest
+import numpy as np
+from pytest import approx
+
+"""
+Format of Reference Dicts is based upon return values from google.com of the 
+form "1s in ns". Canonical values used where clealy available (e.g. seconds in
+minute/hour). All values in each dict should be equivalent quantities after the
+conversions are performed. This reflects how we would naturally form the 
+equalities that we would use to represent units.
+That is, we would represent the equivalency between m & cm as follows:
+lengths['m'] == lengths['cm']
+         1   ==         100         
+"""
+
+# ns us ms s min hr day year
+times = {'s': 1,
+         'ns': 1e9,
+         'us': 1e6,
+         'ms': 1e3,
+         'min': 1/60,
+         'hr': 1/(60*60),
+         'day': 1/(60*60*24),
+         'year': 1/(60*60*24*365)
+         }
+@pytest.fixture(params=list(times.keys()))
+def timepair(request):
+    return {'s': times['s'], request.param: times[request.param]}
+
+
+# km m cm mm um nm A in nmi ft yd mile mi
+lengths = {'m': 1,
+           'km': 1e-3,
+           'cm': 1e2,
+           'mm': 1e3,
+           'um': 1e6,
+           'nm': 1e9,
+           'A': 1e10,
+           'in': 39.3701, #1/0.0254,
+           'nmi': 1/1852,
+           'ft': 39.3701/12, #1/0.0254/12,
+           'yd': 39.3701/36, #1/0.0254/12/3,
+           'mile': 0.000621371, #1/0.0254/12/5280,
+           'mi': 0.000621371, #1/0.0254/12/5280,
+           }
+@pytest.fixture(params=list(lengths.keys()))
+def lengthpair(request):
+    return {'m': lengths['m'], request.param: lengths[request.param]}
+
+
+# kg g mg lbm lb oz slug u amu
+masses = {'kg': 1,
+          'g': 1e3,
+          'mg': 1e6,
+          'lbm': 2.204623,
+          'lb': 2.204623,
+          'oz': 2.204623*16,
+          'slug': 0.06852177,
+          'u': pm.units.const_Na * 1000,
+          'amu': pm.units.const_Na * 1000
+          }
+@pytest.fixture(params=list(masses.keys()))
+def masspair(request):
+    return {'kg': masses['kg'], request.param: masses[request.param]}
+
+
+# N kN lb kgf lbf oz
+forces = {'N': 1,
+          'kN': 1e-3,
+          'lb': 0.224809,
+          'kgf': 1/pm.units.const_g,
+          'lbf': 0.224809,
+          'oz': 0.224809*16,
+          }
+@pytest.fixture(params=list(forces.keys()))
+def forcepair(request):
+    return {'N': forces['N'], request.param: forces[request.param]}
+
+
+# kmol mol lbmol n Nm3 Ncum NL Ncc scf sci
+molars = {'mol': 1,
+          'kmol': 1e-3,
+          'lbmol': 0.0022046226,
+          'n': 1 * pm.units.const_Na,
+          'Nm3': 1 / pm.units.const_dstd,
+          'Ncum': 1 / pm.units.const_dstd,  # meters cubed occupied by 1 mol
+          'NL': 1 / pm.units.const_dstd * 1000,
+          'Ncc': 1 / pm.units.const_dstd * 1e6,
+          'scf': 1 / pm.units.const_dstd * 35.3147,
+          'sci': 1 / pm.units.const_dstd * 61023.7,
+          }
+@pytest.fixture(params=list(molars.keys()))
+def molarpair(request):
[…]
