### 2022-10-10 dda2178
Correct inplace bugs that occur when passing inplace with a type that isn't a numpy array, and thus isn't mutable.
