### 2022-10-10 a23832f
Cleaned up the test code now that the inplace issue is fixed.
