### 2022-10-10 597ef52
Corrected the documentation of how the conversion table is built and the example listed.
