### 2022-10-10 chmarti1/PYroMat#61 @jranalli
> Furthermore, the example conversion inch to feet is not correct:
>
> https://github.com/chmarti1/PYroMat/blob/d65cb2c73bd14b756deca4b7271246322515d221/src/pyromat/units.py#L92-L94

Oh I see what the concern is here. It's that the example dict listed is incorrect.

The definition is:
```python
conv = self.table[from_units] / self.table[to_units]
```

So if `from_units` are inches and `to_units` are ft, we need in = 1/12 and ft = 1 in the table. It's actually an inverse of how the documentation describes it. Added a commit to the pull request #63 clearing up the documentation.
