### 2022-10-10 chmarti1/PYroMat#60 @jranalli
Fixed the text comment for the paper by deleting the end of the sentence.
