### 2022-10-10 chmarti1/PYroMat#61 @chmarti1
Confirmed.  It looks like the in-line documentation has it backwards.  See line 147 in `units.py`, the conversion factor is calculated as

https://github.com/chmarti1/PYroMat/blob/d65cb2c73bd14b756deca4b7271246322515d221/src/pyromat/units.py#L147

I seem to remember making a decision to reverse this behavior back when the units module was under development.  It looks like the documentation was not correctly updated.  Good catch.
