### 2022-10-19 #4757 @chmarti1
God question.  `mp1` uses a generalized "Span and Wagner" equation of state that applies over both liquid and vapor phases without piece-wise transitions.  If you are curious, the `mp1` model is developed in some detail in chapter 7 (pg 97-114) of the [PYroMat handbook](http://pyromat.org/pdf/handbook.pdf).

The `if97` class uses the [IF97](http://iapws.org/relguide/IF97-Rev.html) piece-wise formulation that you probably know already.  The IAPWS actually published an earlier model in 1995 using the Span and Wagner formulation, and it was [updated](http://iapws.org/relguide/IAPWS-95.html) in 2018.

The IF97 is handy for engineering use, because it includes polynomial expansions good for "inverse" property calculations, so it was the first thing I tried.  If I had to write a code to model a Rankine cycle from scratch, I'd use IF97.  However, Span and Wagner equations of state are now almost universal in applications like PYroMat because:
- Piece-wise approaches are heavily dependent on the substance being modeled.  The regions' number, their arrangement, and even their formulation is so variable, it makes writing a generalized class a daunting task.
- Piece-wise approaches require the algorithm to determine the region of the state prior before calculating properties.  If the primary properties are known, that is simple, but in any other case, even determining the region can require iteration before any properties are calculated.
- Numerous models using the Span and Wagner approach are now available in the literature, so extensibility to multiple substances is not a problem.
- For super-critical, vapor, and mixture states, (temperature, density) is a numerically superior means for determining the state.  It addresses almost all of the singularities that occur near phase changes and the critical point, and it even permits the calculation of meta-stable properties.  PYroMat does not currently support meta-stable states, but the underlying models are theoretically capable of estimating their properties.
- Once the potential numerical convergence problems are addressed, the numerical iteration required to specify states by (T,p) when the properties are calculated from (T,d) is not very expensive.  Really, this problem has to be solved with IF97 anyway, because it uses (T,d) in the super-critical region.
