### 2022-10-24 #4757 @fwitte
@chmarti1, thank you very much for the detailed explanation. I was wondering, why I could not replicate some of the bugs in IF97, which seem to exist in both, [CoolProp](https://github.com/CoolProp/IF97) and the [IAPWS package](https://github.com/jjgomera/iapws), also see https://github.com/CoolProp/CoolProp/issues/1918.
