### 2022-10-24 chmarti1/PYroMat#65 @fwitte
[Review] Repo/Docs are missing some guidelines for contribution
Especially for less experienced git users and new developers, who want to contribute to your project, it would be very helpful to have short section in the GitHub repo and/or the documentation with information on

- [ ] how and where (e.g. have some module overview) to make changes in the source code and open pull requests,
- [ ] how to run the software tests (kind of related to https://github.com/chmarti1/PYroMat/issues/66) and
- [ ] code style and code documentation guidelines.
