### 2022-10-24 chmarti1/PYroMat#66 @fwitte
[Review] Include `extra_requires` in setup.py
Something like

```
...
    extras_require={
        'dev': [
            'pytest',
        ]
    },
...
```

is very useful in my opinion, so people can install `pip install pyromat[dev]` or `pip install -e .[dev]` if they want to have all requirements for the development process as well.
