### 2022-10-24 #4757 @fwitte
@jranalli, @chmarti1: Thank you for your replies and in depth explanation. Here are a few lines for my review:

# Summary

Overall, I find this a very useful package. To my knowledge, the NASA polynomes are not included in other open source packages like [thermo]( https://github.com/CalebBell/thermo), [CoolProp](https://github.com/CoolProp/IF97) and [iapws](https://github.com/jjgomera/iapws), but only in EES or other proprietary software. Especially the extensive handbook is well written and contains a lot of information helping both, people learning thermodynamics and those, who have been working with unit based simulation but need a refresher in fluid property modeling.

The online documentation is also well written and features some small gimmicks with the live apps, giving potential users a quick idea, what the software does. I have some suggestions for the online documentation, which I listed under ideas/suggestions, which are not mandatory for the paper to be accepted.

# Major Issues

All major issues have been fixed already, see the comments of `@espottesmith`

# Minor Issues

- https://github.com/chmarti1/pyrodoc/issues/4
- https://github.com/chmarti1/PYroMat/issues/60
- https://github.com/chmarti1/PYroMat/issues/61 (already fixed in open PR)
- https://github.com/chmarti1/PYroMat/issues/62
- https://github.com/chmarti1/PYroMat/issues/65
- https://github.com/chmarti1/PYroMat/issues/66

# Ideas and Suggestions for future development

- Make an API to let the user build his/her own mixtures: While you can just copy a `.hpd` file, give it a new id and some mixture contents, this might be hard to find for less experienced users and you are limited to a somewhat fixed mixture composition. It would be cool, if the user could build his or her own mixtures by providing the molar or mass based mixture contents and it does not seem too hard to implement.
- I often prefer an non live version of the help, which I can search through while developing code. In my opinion, it is very helpful if classes and and methods are written somewhere for people who want to include your package within their software. There are currently some ideas for reworking the fluid property back end of my main package, and I am very much considering adding your package as option there.
- Finally, a search function letting users search the online documentation for api, examples, explanations, ... would be a good feature. (I have about 100 searches per month on my main package.)

I am not sure, how complicated it is to include the last two features in your documentation. I do very much like using sphinx (with readthedocs if you do not want to self-host) as it is very easy and convienient to include these.

# Conclusing

Recommend accepting the paper as is after the last two issues have been resolved.
