### 2022-10-24 9b49e88
Add extras_require config for pytest in dev mode
