### 2022-10-24 #4757 @jranalli
Thanks for your work on the review.

- Added a pull request for the `extra_requires`
- Discussing the guidelines for contribution with @chmarti1

Definitely appreciate the suggestions and we'll consider those as we move forward.
