### 2022-10-26 chmarti1/PYroMat#65 @chmarti1
I don't object to the `extra_requires` idea.  That's a good suggestion.  I've modified the README to encourage people to contribute.  In terms of programming / style guide, the handbook is going to need to suffice for now.  Community contributions have been limited to small snippets of code so far, so it hasn't been important to develop a guide.
