### 2022-10-26 073e394
Updated readme to include recommendations for community involvement.
