### 2022-10-26 0dc73a5
Remove legacy tests. Add test_units to github workflow. Closes #69
