### 2022-10-28 def313f
Add the DOI to README
Added the Zenodo DOI badge to the README
