### 2022-10-28 #4757 @jranalli
Ok so I don't know how to mark the checkboxes, but I have double checked that everybody's ORCID and affiliation appears correctly on the paper, the version release number is 2.2.4 and it is uploaded to Zenodo with a DOI of https://doi.org/10.5281/zenodo.7262173
