### 2022-11-02 #4757 @jranalli
He has been a contributor to the overall project for years but hasn't made code contributions and so doesn't show on Github. We looked at manually adding an author to the Zenodo archive, but it didn't look like there was an easy way to do so. We can take a look at whether there's a different method to do so, unless you have some suggestion that we didn't think of?
