### 2022-11-02 #4757 @danielskatz
It's very easy to add an author to Zenodo. See the section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org for how to do this.
