### 2022-11-02 #4757 @fwitte
@jranalli: Usually, you can go to zenodo, click on edit publication (only the person who uploaded that publication) and then enter the names of the researches. Looks like so:

![grafik](https://user-images.githubusercontent.com/20252362/199559559-5760f1d8-3ab9-480c-be32-171688be192a.png)
