### 2022-11-03 #4757 @jgostick
Hi @jranalli, the references in the pdf need some work. In addition to the issues pointed out above, there are many references missing their doi.  3 of the last 4 in the list have dois at the bottom of the page: https://doi.org/10.1023/A:1022362231796.  Not sure about the NASA one, but can you take a look?
