### 2022-11-03 #4757 @jgostick
Also, for the nist paper, what about: https://doi.org/10.1021/acs.iecr.2c01427 instead of just linking to the nist website
