### 2022-11-03 8fe9a09
Updating some reference DOIs
