+++ joss2022/paper.bib
@@ -2,7 +2,7 @@
 	address = {Columbus, OH},
 	title = {Problem-based {Learning} {Module} for {Teaching} {Thermodynamic} {Cycle} {Analysis} using {PYroMat}},
 	url = {https://peer.asee.org/problem-based-learning-module-for-teaching-thermodynamic-cycle-analysis-using-pyromat},
-	doi = {https://peer.asee.org/28757},
+	doi = {10.18260/1-2--28757},
 	urldate = {2019-02-14},
 	author = {Martin, Christopher Reed and Ranalli, Joseph and Moore, Jacob Preston},
 	month = jun,
@@ -46,13 +46,19 @@
 	pages = {2498--2508},
 }
 
-@misc{Lemmon:2018,
-	title = {{NIST} {Standard} {Reference} {Database} 23: {Reference} {Fluid} {Thermodynamic} and {Transport} {Properties}-{REFPROP}},
-	url = {https://www.nist.gov/srd/refprop},
-	publisher = {National Institute of Standards and Technology},
-	author = {{E.W. Lemmon} and {Ian H. Bell} and {M.L. Huber} and {M.O. McLinden}},
-	year = {2018},
+@article{Huber:2022,
+author = {Huber, Marcia L. and Lemmon, Eric W. and Bell, Ian H. and McLinden, Mark O.},
+title = {The NIST REFPROP Database for Highly Accurate Properties of Industrially Important Fluids},
+journal = {Industrial \& Engineering Chemistry Research},
+volume = {61},
+number = {42},
+pages = {15449-15472},
+year = {2022},
+doi = {10.1021/acs.iecr.2c01427},
+URL = {https://doi.org/10.1021/acs.iecr.2c01427},
+eprint = {https://doi.org/10.1021/acs.iecr.2c01427},
 }
+
 
 @misc{pmhandbook,
 	title = {The PYroMat User and Developer Handbook},
@@ -70,6 +76,7 @@
   number = {1},
   year = {2003},
   pages = {1-39},
+  doi = {10.1023/A:1022390430888},
 }
 @ARTICLE{span:2003:2,
   author = {R. Span and W. Wagner},
@@ -79,6 +86,7 @@
   number = {1},
   year = {2003},
   pages = {41-109},
+  doi = {10.1023/A:1022310214958},
 }
 @ARTICLE{span:2003:3,
   author = {R. Span and W. Wagner},
@@ -88,11 +96,18 @@
   number = {1},
   year = {2003},
   pages = {111-162},
+  doi = {10.1023/A:1022362231796},
 }
-@MISC{nist:webbook,
-  howpublished = {National Institute for Standards and Technology, https://webbook.nist.gov/},
-  note = {accessed: Jun, 2021},
+@book{nist:webbook,
+  title = {NIST Chemistry WebBook, NIST Standard Reference Database Number 69},
+  author = {National Institute of Standards and Technology},
+  editor = {P.J. Linstrom and W.G. Mallard},
+  publisher = {National Institute of Standards and Technology},
+  address = {Gaithersburg MD, 20899},
+  year = {2022},
+  doi = {10.18434/T4D303},
 }
+
 @TECHREPORT{mcbride:1993,
   author = {Bonnie J. {McBride} and Sanford Gordon and Martin A. Reno},
   title = {Coefficients for Calculating Thermodynamic and Transport Properties of Individual Species},
