### 2022-11-03 e3aa8c8
Updating some reference DOIs
