+++ joss2022/paper.bib
@@ -100,7 +100,6 @@ eprint = {https://doi.org/10.1021/acs.iecr.2c01427},
 }
 @book{nist:webbook,
   title = {NIST Chemistry WebBook, NIST Standard Reference Database Number 69},
-  author = {National Institute of Standards and Technology},
   editor = {P.J. Linstrom and W.G. Mallard},
   publisher = {National Institute of Standards and Technology},
   address = {Gaithersburg MD, 20899},
