### 2022-11-03 #4757 @jranalli
> 10.18260/1-232083 is INVALID
This one appears to resolve correctly in the paper, and the [link](https://doi.org/10.18260/1-2--32083) works. It seems like the `--` is parsing incorrectly in editorialbot's tests?
![image](https://user-images.githubusercontent.com/7864460/199768402-1c4af763-8703-407d-b224-894f6374d916.png)

That will likely be a similar issue with the (now fixed) peer.asee.org paper, as it has the same string in its DOI.

For the REFPROP NIST reference, thanks for the link you suggested. For the second NIST ref though I found a DOI for the whole webbook, which is more appropriate for the Shomate equation.

I looked in depth for the NASA paper, but based on its age, I don't think a DOI exists. My understanding is that DOIs were only instituted around 2000.

So the only two things without DOIs right now are the Pyromat reference handbook and the NASA paper.
