### 2022-11-04 #4757 @jgostick
Yes, there does seem to be a bug in joss's doi parser. I can confirm that the '--' are missing from the editorialbot output and if I put them back in, the doi resolves just fine.  I'll figure how to report this to the team.
