### 2022-11-04 #4757 @jgostick
I opened an issue [here](https://github.com/openjournals/joss/issues/1156).  I think for now we can proceed with the process.
