### 2022-11-04 #4757 @xuanxu
Two hyphens is a LaTeX command for a en-dash, so the parser creates an invalid DOI.
I've opened [a PR](https://github.com/chmarti1/PYroMat/pull/74) that should fix it.
