### 2022-11-04 #4757 @jranalli
@xuanxu I committed that change, but it messes up the printout in the paper, so I need to revert it:
![image](https://user-images.githubusercontent.com/7864460/199980541-cb1063af-8cab-4d3d-8305-f56308fbb608.png)

I think the problem is just with the bot's parsing when it does the tests for valid DOIs, rather than the document rendering. The document rendered fine before, the only issue was the output from the `check references` command.
