### 2022-11-04 accc97d
Revert "Encode double hyphens"
