+++ joss2022/paper.bib
@@ -2,7 +2,7 @@
 	address = {Columbus, OH},
 	title = {Problem-based {Learning} {Module} for {Teaching} {Thermodynamic} {Cycle} {Analysis} using {PYroMat}},
 	url = {https://peer.asee.org/problem-based-learning-module-for-teaching-thermodynamic-cycle-analysis-using-pyromat},
-	doi = {10.18260/1-2{-}{-}28757},
+	doi = {10.18260/1-2--28757},
 	urldate = {2019-02-14},
 	author = {Martin, Christopher Reed and Ranalli, Joseph and Moore, Jacob Preston},
 	month = jun,
@@ -26,7 +26,7 @@
 @inproceedings{Ranalli:2019,
 	address = {Tampa, FL},
 	title = {An {Online} {Tool} for {Facilitating} {Thermodynamic} {Property} {Lookups}},
-	doi = {10.18260/1-2{-}{-}32083},
+	doi = {10.18260/1-2--32083},
 	publisher = {American Society for Engineering Education},
 	author = {{Joseph Ranalli} and {Christopher Martin} and {Jacob Moore}},
 	month = jun,
