### 2022-11-06 #4757 @jgostick
Hi @openjournals/pe-eics, note that the two INVALID DOIs in the reference check are false alarms.  The dois in the paper have 2 dashes, and this is breaking editorialbot when it attempts to check them.
