### 2022-11-06 729c3d5
Fixes typos in JOSS paper
