+++ joss2022/paper.md
@@ -31,22 +31,22 @@ bibliography: paper.bib
 
 # Summary
 
-Lookup of thermodynamic properties of substances is a common task in numerous engineering disciplines and is also an important skill taught to students in several common engineering disciplines. While properties were traditionally found using tables of thermodynamic data, availability of modern software has essentially eliminated the need for laborious use of tables, and has created opportunities to support quick iteration and revision of calculations. As thermodyanmic data are published in a variety of formats, it is necessary for software to provide a standardized interface to operate on a number of different data source formats. Additionally, thermodynamic software needs to be numerically efficient and robust, to deal with the diverse computational methodologies required to produce property calculations. PYroMat seeks to create a tool that meets these needs and can grow with users at different levels of expertise. It offers a low barrier to entry and is easy enough for students to use, but also powerful and flexible enough to serve the needs of engineering professionals and scientific researchers.
+Lookup of thermodynamic properties of substances is a common task in numerous engineering disciplines and is also an important skill taught to students in several common engineering disciplines. While properties were traditionally found using tables of thermodynamic data, availability of modern software has essentially eliminated the need for laborious use of tables, and has created opportunities to support quick iteration and revision of calculations. As thermodynamic data are published in a variety of formats, it is necessary for software to provide a standardized interface to operate on a number of different data source formats. Additionally, thermodynamic software needs to be numerically efficient and robust, to deal with the diverse computational methodologies required to produce property calculations. PYroMat seeks to create a tool that meets these needs and can grow with users at different levels of expertise. It offers a low barrier to entry and is easy enough for students to use, but also powerful and flexible enough to serve the needs of engineering professionals and scientific researchers.
 
 # Statement of need
 
-As far back as the birth of the industrial revolution, engineers and scientists have needed precise calculations for the thermodynamic properties of fluids to predict the behaviors of systems of global importance.  Today, people working in aerospace propulsion, electrical power generation, plasma physics, refrigeration, building heating and cooling, combusion, and countless other fields still rely on decades of excellent data for these calculations.  It is exceedingly rare to find either the original data or software that performs these important calculations in the public domain.
+As far back as the birth of the industrial revolution, engineers and scientists have needed precise calculations for the thermodynamic properties of fluids to predict the behaviors of systems of global importance.  Today, people working in aerospace propulsion, electrical power generation, plasma physics, refrigeration, building heating and cooling, combustion, and countless other fields still rely on decades of excellent data for these calculations.  It is exceedingly rare to find either the original data or software that performs these important calculations in the public domain.
 
-The current industry standard, NIST's `REFPROP` [@Huber:2022], is neither free nor open.  Another excellent alternative is `coolprop`[@Bell:2014], which is an NSF-funded flexible interface for an impressive variety of languages, which focuses heavily on multi-phase substances.  There are also other less widely embraced alternatives, but `PYroMat` distinguishes itself because:
+The current industry standard, NIST's `REFPROP` [@Huber:2022], is neither free nor open.  Another excellent alternative is `coolprop` [@Bell:2014], which is an NSF-funded flexible interface for an impressive variety of languages, which focuses heavily on multi-phase substances.  There are also other less widely embraced alternatives, but `PYroMat` distinguishes itself because:
 a) every aspect of it (and its dependencies) is open source, and 
 b) it provides a standard Pythonic interface simultaneously simple enough for novices and powerful enough for professionals. 
 Several conference publications have already described early versions of the software
[…]
