### 2022-11-06 #4757 @kyleniemeyer
> Hi @openjournals/pe-eics, note that the two INVALID DOIs in the reference check are false alarms. The dois in the paper have 2 dashes, and this is breaking editorialbot when it attempts to check them.

Thanks for the heads-up @jgostick—however, I did notice that some information was missing from those entries (notably the proceedings/book title).

@jranalli I made a few typo fixes and bibliography updates in https://github.com/chmarti1/PYroMat/pull/76, could you merge that?
